+++ Studenda.Core.Test.Integration/Studenda.Core.Test.Integration.csproj
@@ -8,7 +8,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Studenda.Model\Studenda.Model.csproj" />
+      <ProjectReference Include="..\Studenda.Core\Studenda.Core.csproj" />
     </ItemGroup>
 
 </Project>
