+++ Studenda.Core/Model/Account/User.cs
@@ -1,8 +1,8 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Studenda.Model.Data.Configuration;
-using Studenda.Model.Shared.Link;
+using Studenda.Core.Data.Configuration;
+using Studenda.Core.Model.Link;
 
-namespace Studenda.Model.Shared.Account;
+namespace Studenda.Core.Model.Account;
 
 /// <summary>
 /// Пользователь.
