+++ Studenda.Core/Model/Entity.cs
@@ -1,8 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Studenda.Model.Data.Configuration;
+using Studenda.Core.Data.Configuration;
 
-namespace Studenda.Model.Shared;
+namespace Studenda.Core.Model;
 
 /// <summary>
 /// Модель стандартного объекта с соответствующей
