+++ Studenda.Core/Studenda.Core.csproj
